refactor(kftray-kube): split stop.rs by domain
Split kube/stop.rs (784 LOC) into kube/stop/:
- mod.rs: re-exports
- address.rs: helper service IPC for address release
- cleanup.rs: load_configs + delete_proxy_cluster_resources
- single.rs: stop_port_forward / stop_port_forward_with_mode + tests
- all.rs: stop_all_port_forward / stop_all_port_forward_with_mode + tests

Tests travel with their code. Zero behaviour change.
