+++ crates/kftray-kube/src/kube/stop/mod.rs
@@ -1,0 +1,14 @@
+mod address;
+mod all;
+pub(crate) mod cleanup;
+mod single;
+
+pub use all::{
+    stop_all_port_forward,
+    stop_all_port_forward_with_mode,
+};
+pub(crate) use cleanup::delete_proxy_cluster_resources;
+pub use single::{
+    stop_port_forward,
+    stop_port_forward_with_mode,
+};
